+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Recipe Provider for standard and breeding recipes
 - Language provider for the English lang
 - Coop to generate drops from spawn eggs
+- Random layTime to ChickenData
 
 ### Changed
 - Change build to 1.19.2